Addred Permission and Seeder for provider, category and services

+++ app/Http/Controllers/CategoryController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use Inertia\Inertia;
 use App\Models\Category;
 use Illuminate\Http\Request;
-use Inertia\Inertia;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -13,6 +14,10 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->hasPermission('category.view')) {
+            abort(403);
+        }
+        
         $categories = Category::with('children')->get();
         return Inertia::render('Backend/Categories/Index', ['categories' => $categories]);
     }
@@ -22,6 +27,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->hasPermission('category.create')) {
+            abort(403);
+        }
+
         $categories = Category::select('id', 'category_name as label')->get();
         return Inertia::render('Backend/Categories/Edit', ['categories' => $categories]);
     }
@@ -31,6 +40,10 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Auth::user()->hasPermission('category.create')) {
+            abort(403);
+        }
+
         $request->validate([
             'category_name' => 'required|string|max:255',
             'parent_id' => 'nullable|exists:categories,id',
@@ -46,7 +59,7 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
-        return Inertia::render('Backend/Categories/Edit', ['category'=> $category]);
+        //
     }
 
     /**
@@ -55,6 +68,10 @@ class CategoryController extends Controller
 
     public function edit($id)
     {
+        if (!Auth::user()->hasPermission('category.update')) {
+            abort(403);
+        }
+
         $category = Category::findOrFail($id);
         $categories = Category::select('id', 'category_name as label')->get();
         return Inertia::render('Backend/Categories/Edit', ['category' => $category, 'categories' => $categories]);
@@ -66,6 +83,10 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        if (!Auth::user()->hasPermission('category.update')) {
+            abort(403);
+        }
+
         $request->validate([
             'category_name' => 'required|string|max:255',
             'parent_id' => 'nullable|exists:categories,id',
@@ -82,6 +103,10 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::user()->hasPermission('category.delete')) {
+            abort(403);
+        }
+
         $category = Category::findOrFail($id);
         $category->delete();
         return redirect()->route('categories.index')->with('success', 'Category deleted successfully.');
